Adding Errors for Non-Static Instance Variables and Methods Inside of Abstract Final Classes for FFP
Summary:
Making the FFP more robust by allowing the parser to detect errors when non-static instance variables or methods are wrongly included in an abstract final class. Currently, hhvm does produce an error for this case but the FFP and the Hack parser don't.

This new error detection was implemented by adding new code, as well as modifying preexisting code (namely the extract_abstract_keyword method) for a more flexible but equally as effective and succinct code base.

Reviewed By: ericlippert

Differential Revision: D5787779

fbshipit-source-id: 4530f559bdb0a4885abcaefabf1a93211d3748ef