Show relevant constraints on type parameters in hover types
Summary:
Currently the `hh_show` pseudo-function and infer-at-pos service display the type
of a local variable. But the `instanceof` feature can introduce constraints (`as`,
`super` and equality constraints) on type parameters, and these are currently not
shown, which can make debugging type errors difficult. These constraints are now
displayed for any type parameters occurring in the type.

Reviewed By: jamesjwu

Differential Revision: D5804829

fbshipit-source-id: 805039505276f24edf31069dfc1f48c729f78f32