Refactor coroutine rewriters
Summary: I'm making a series of small refactorings that will end up with outer variables being captured correctly by closures. In this refactoring: the "context" object now holds the original node of the coroutine, so it can be used to rewrite the coroutine body. We no longer tramp around the coroutine body redundantly.

Reviewed By: michaeltingley

Differential Revision: D5803869

fbshipit-source-id: b4a01c73f7a7cde5ed2d0129991af5dd9ee454d5