Compute outer variables in coroutine context constructor
Summary:
We're going to need to know what the outer variables are of a lambda in several places in the coroutine lowering code. Rather than tramp them around, let's just compute them up front and put them in the coroutine context. We already tramp the context around everywhere.

This has no semantic changes; it just moves a computation to earlier in the pipeline and caches the results. In the next few diffs I'll change how that information is used, and fix a bug in the outer variable computation.

Reviewed By: michaeltingley

Differential Revision: D5804156

fbshipit-source-id: 42610c95aadf660819a2efbb434a0a21d4dd6fde