Fix a bug in dce
Summary:
When we do a PopInputs, or PopAndReplace, if the instructions that
generated the inputs are also dce-able we'll add a Kill for the
Pop*. commitActions will then convert Kill + PopInputs to kill, and
Kill + PopAndReplace to Replace.

But combineActions only works within a single block. When we combined
the actions from multiple blocks, we just merged the actions, with the
first in taking precedence. The result was that we'd kill the
instructions that generated the inputs, but still try to pop them.

I can't come up with a repro as things stand, but another diff I'm
working on revealed the issue.

Reviewed By: ricklavoie

Differential Revision: D5806765

fbshipit-source-id: 27b848cdfdbaa0cd140c0fdfac8eb19646ae7a66