Correctly handle illegal FPushObjMethodDs
Summary:
If early on we see that the object passed to FPushObjMethodD is of
type ?X, and we know that X::foo returns an int, we'll infer that
either the FPush throws, or the result of the corresponding call is
int.

But later on we might prove that the object is actually
Null. Currently we say we know nothing about the result of the FCall,
and it reverts to TInitGen.

This fixes it to make the next bytecode unreachable, so that types
remain consistent.

Reviewed By: ricklavoie

Differential Revision: D5815538

fbshipit-source-id: 05431f82a55107e7b6790d12179a4e8a1de8816d