Send correct exception name when hitting an exception breakpoint
Summary:
Our XDebug extension is sending the exception name set in the exception breakpoint, rather than the exception that was hit in PHP. This means the "break on any exception" semantics, which set the exception name to "*" are always returning "*" to the debugger front end as the exception that was hit.

Fixing this will pave the way for exception breakpoint filtering on Exn type.

Reviewed By: mostafaeweda

Differential Revision: D5826126

fbshipit-source-id: c33a8acaa9ce6748e46494551d1f3070164f9308