Fix computation of outer variables involving anonymous methods
Summary:
For lambdas we deduce which local variables are captured outer variables of the lambda; for anonymous functions we list them in a "use" clause. We therefore can -- and must -- stop looking for outer variables when we hit an anonymous method.

The outer variable computation should now be correct, but we still don't copy the outer variables around correctly when lowering a coroutine. That'll be the next diff.

 

Reviewed By: michaeltingley

Differential Revision: D5813792

fbshipit-source-id: 3a23796adccc8a9cb231b1eae0ac74fefff3348f