Improve FF parsing of collection initializer expressions
Summary:
Collection initializers have possible forms

    Map{"a"=>1, "b"=>2}
    Vector{2, 4, 6}
    Pair{$foo, $bar}

Whether the list items are key-value pairs or not depends on the type of the collection.

We used to check the first one and see if the remainder matched, and give an obscure error with bad error recovery if you had a mismatch.

That's not what the original parser did, and it's not what the FF parser should do either. We should rather parse this as a list of (expressions or pairs) and not a list of expressions or a list of pairs.

The type checker then gives the right error if there is a pair in a non-map, and so on.

Reviewed By: vladima

Differential Revision: D5823505

fbshipit-source-id: f91905f290acf5c785f5cd878f13ca63754d31b7