Keep track of namespace for closures
Summary: Closures defined within functions or methods hadn't keep track of the namespace prevailing at their definition. This made it impossible to pass the correct (unqualified) fallback function name in function calls in the lambda body. Here we return a map of closures to their namespaces to pass to the emitter phase.

Differential Revision: D5773174

fbshipit-source-id: 9df788bf4e89b934cd9cbd5a2868fc8759e6bad3