Fix closure contexts
Summary:
Contexts for closures need to be fixed up so that they point to the
containing class, not the closure itself.

This was happening before analyzing the function body, but (in some
cases) after the CollectedInfo had been created.

During the analysis phase, that doesn't matter, but during the final
optimization phase the CollectedInfo constructor uses the class to
retrieve property information from the index. This meant that we
created a suboptimal FuncAnalysis object to pass into optimize_func;
but the FuncAnalysis itself had the correct context, so optimize_func
*did* get to use the information from the index, which could result in
inconsistencies between what optimize_func believed, and what the
FuncAnalysis believed.

Reviewed By: ricklavoie

Differential Revision: D5834003

fbshipit-source-id: bea6c77b05e5d7e3a1d7edf94ca10b5194003b9b