Fix PHP debugger showing resolved breakpoints as unresolved
Summary: HHVM is sending a breakpoint resolved message only when a PHP file is loaded (it checks unresolved breakpoints against compilation units in the new file).  If setting a new breakpoint in a file that is already loaded, it is marked resolved in XDebug but no notification is sent to the debugger client, meaning the UX is showing the bp as unresolved.

Reviewed By: mostafaeweda

Differential Revision: D5829612

fbshipit-source-id: a8f8eb30a5506a483d265f4692b33cf1760dd98c