Add tests from deps + bugfixes
Summary:
This adds all the unit tests from the hphp/hack/test/deps folder to test and compare against, and also fixes bugs associated with these tests. It also changes --dump-bazooka-deps to be more consistent with what we expect the toplevel dependency analyzer to output(by getting rid of FunName and GConstName and instead only returning Fun and GConst). I couldn't decide whether to just always add these dependencies or not, but since we're going to be changing how they work anyway, I figured it'd be better to be simpler.

The only fix included in this diff for the dependency analyzer is to always add both the current namespace and the global namespace when adding dependencies for constants and functions.

Reviewed By: dabek

Differential Revision: D5739964

fbshipit-source-id: 119459cbe51de4c73b81849a036d5610999ae336