only send an exception message once to client
Summary:
What happens when clientLsp raises an exception?

At first... if handling a request then it responded with an error response to the client, otherwise it silently swalled the exception.

After D562714... if handling a request then it responded with an error response to the client. In addition, in all cases, it also sent a window/telemetryEvent notification with the error.

[This broke one of the integration tests, which was expecting to see just the exception inside the error-response, but now also got a telemetryEvent with the exception as well. Also, I think it's ugly for the client to receive information about the same exception twice]

After this diff... if handling a request then it responds with an error response to the client, otherwise it sends a window/telemetryEvent notification with the error.

I think this gives the best of both worlds\! And it makes for easier-to-read client logs.

Reviewed By: arxanas

Differential Revision: D5829732

fbshipit-source-id: 301becdef2753aca87c37b847080f1d3086924d9