don't send two responses to "initialize"
Summary:
What happens if you start lsp but hh_server isn't yet running? In D5620642 we changed it so when we receiving an "initialize" message, and we fail to connect, then we straight away go into a lost state.

There was a bug in that diff which resulted in two JsonRPC responses being sent in response to an "initialize" request, one claiming success, one claiming failure.

In the diff, the job of the "initialize" handler inside handle_event is explained by the comment:

  (* We always succeed in initialize and send a response here. *)

The actual work of attempting to connect is left to a different method, "connect_from_preinit_if_necessary". That method
attempts to connect to hh_server. If it fails (as indicated by an exception) then it returned that exception to be thrown.
But the exception handler decided that the exception was thrown in the context of a request, so it should give back an error-response. This is incorrect.

I changed it so the exception is reported as a JsonRPC notification, not as a JsonRPC response.

Reviewed By: arxanas

Differential Revision: D5829782

fbshipit-source-id: 50b9bb065018032e3305da6d3a3ecac766058a36