quick-fix for integration tests
Summary:
In D5620649 we moved clientLsp to never autostart. The assumption is that hh_server should always be running on a dev machine. If it's not, then clientLsp will report that to the user and offer a handy "restart" button.

The assumption doesn't hold for test projects\! So, this diff changes the test runner to start hh_server (and wait until it's ready) before starting the test.

Also,

In D5620642 we moved clientLsp to connect asynchronously to hh_server: the lsp "initialize" method will return immediately with success, and it will immediately *start* a connection to hh_server, but it won't necessarily finish.

How should the unit-tests deal with this? If they send requests like "definition" before the connection is finished, they'll just get back an error about ServerBusy.

This diff contains a quick fix: after sending initialize, we'll just wait 0.5 seconds. By that time (and thanks to the fact that we know hh_server was already responsive before we kicked off clientLsp), we can safeuly assume that the connection has been made.

Ultimately it'd make sense to send some kind of proper notification from the server to the client to say when its connection status has changed. Or maybe the client could poll with an "are-you-connected" request. Or maybe the client could send a long-lived "please fulfill this request once you are ready". I'm not sure what is the best design. (And the answer that's easiest to consume from Nuclide is different from what's easiest to consume from test_lsp.py). So I'll hold off on the proper fix for now.

Differential Revision: D5829797

fbshipit-source-id: 3bbaeac002e442bbbb570abd31399819c3b89ddb