Bug-patch; drop XHP attributes if any name contains a colon
Summary:
Unfortunately, `Parser_hack` has a bug handling XHP attribute names containing a colon; it drops everything after on the floor. Concretely
```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

function f(): :xhp {
    /* HH_IGNORE_ERROR[1002] */
    <use:foo xlink:href="#p" bar="0" baz="qux"/>;
}
```
comes out as
```
(AProgram
 ((Fun
   ((f_mode: Mstrict) (f_tparams: ()) (f_ret_by_ref: false)
    (f_ret: ((p (Happly (p :xhp) ())))) (f_name: (p "\\f")) (f_constrs: ())
    (f_params: ())
    (f_body:
     ((Expr (p (Xml (p :use:foo) (((p xlink) (p (String (p ""))))) ())))))
    (f_user_attributes: ()) (f_fun_kind: FSync)
    (f_namespace:
     ((ns_name: "") (ns_ns_uses: (SMap ())) (ns_class_uses: (SMap ()))
      (ns_fun_uses: (SMap ())) (ns_const_uses: (SMap ()))))
    (f_span: p)))))
```
Notice;
  * The `HH_IGNORE_ERROR` which suppresses the parse error
  * The resulting `Ast` does contain the tag name //with// a colon (`:use:foo`)
  * The resulting `Ast` does //not// contain any of the attributes after `xlink:href`, which it records as `xlink`

This diff is to port that bug into the full fidelity parser, so that we can move fast on other fronts. There is still a slight semantic difference, because `Parser_hack` pushes the lexer to "a `\n` or `}`" whereas the FFP will at least still look for `>` or `/>`.

Reviewed By: ericlippert

Differential Revision: D5834309

fbshipit-source-id: 213ca0b7da3e1a003217ba4ff6357770fa339f91