WIP - remove lexer allocations
Summary:
cut down on unnecessary lexer allocations

the strategy here is to take lexer-producing functions
written in a recursive style with many intermediate lexers
and change them to an iterative style with an int ref keeping
track of the amount to advance.

I'm rewriting the low hanging fruit first ... and avoiding complex mutually recursive functions.

Differential Revision: D5833444

fbshipit-source-id: fcf0b1279d05d27e942a6687ae36e3510a463ba8