Fix some issues with inline analysis and constant folding
Summary:
We can't call anything that inspects the bytecode of another function
from the optimize phase, including analyze_func_inline. This is
guarded by Index::frozen(), but we weren't setting that during the
constant folding pass, and some new code is starting to trigger crashes.

We have heuristics for marking functions and classes as non-persistent
if you apply the class_exists or function_exists builtins directly to
them. analyze_func_inline can cause a lot more of these calls to be
resolved - so disable that heuristic while doing inline analysis.

Reviewed By: ricklavoie

Differential Revision: D5853110

fbshipit-source-id: 9e0d07bf02c30c239dcc1e710cab39f78985df9c