Autostart in case of build id mismatch
Summary:
In an earlier diff we switched over to "never autostart" -- with the expectation that hh_server is always running, and if it's not then something weird has happened, and it's right to ask the user for manual intervention in the form of clicking the Restart button.

But actually there's one (and only one) case where autostart is appropriate: in case of build id mismatch.

* If the server discovers that [version] in .hhconfig has changed, then the server will quit, but the monitor will go ahead and restart it\! ... clientLsp will detect that the persistent connection closed, and will observe that the version has changed, and clientLsp will terminate, and VSCode/Nuclide will relaunch the correct newer version of clientLsp.

* If a newer version of clientLsp or any other hh command attempts to connect to an older version of hh_server (or vice versa), then hh_server terminates and the monitor terminates and the connection attempt raises Exit_with Build_id_mismatch.

In this scenario, the right and proper thing is for the client to do a second connection attempt with {autostart=true}. This will launch the right version of the server. This is the only mechanism we have for launching the right version of the server.

This diff does the right and proper thing.

BUT I'm worried about "version mismatch battles", e.g. if one tool is invoking one version of hh, and another tool (e.g. Nuclide) is invoking a different version, then each will keep terminating the other. In this sort of case we should bail on autostarting, and leave the decision in the hands of the user.

This diff achieves that: it only allows the autostart-after-mismatch once for a given lifetime of clientLsp, and only if there hasn't yet been a successful connection attempt.

Reviewed By: arxanas

Differential Revision: D5850382

fbshipit-source-id: 0eb3b4c01b122b9e8eaa2ef03d5dba9f705ba677