Capture all locals that were declared in outer scopes and don't exist in lambda parameter list
Summary:
We used to have a dedicated logic to figure out which variables should be captured - if variable is definitely assigned inside the lambda then capturing it is unnecessary. However there were a bunch of missing parts in the implementation of definite assignment analysis which lead to bad codegen and also it was not aligned with HHVM behavior leading to spurious failures reported by semdiff.
In this diff decision whether local should be captured is made using pre-collected set of locals from the enclosing scope - local is captured if it is not included in lambda parameter list and is declared in one of enclosing scopes.

Reviewed By: andrewjkennedy

Differential Revision: D5832740

fbshipit-source-id: 4cb3e24e80e076b84afb0a733aab1076591488bf