PHP Debugger: Fix malformed properties defined in object parent class
Summary:
This fixes the long-standing infamous bug in our PHP debugger where private field members of an object that are defined ina base class are displayed in the scopes view with invalid PHP syntax, and are not viewable. Expanding these members in the scopes window used to cause the debugger UI to jump and then fail to display the value. Typing the string in the REPL results in a PHP error due to the invalid syntax.

Consider the following sample code:

  class ParentClass {
    private int $parent_private_f1 = 1;
  }

  class ChildClass extends ParentClass {
    private int $child_private_f1 = 2;
  }

  // BREAKPOINT AT THE FOLLOWING LINE:
  $child = new ChildClass();

When examining $child in the scopes pane, HHVM is returning something like this:
$child = {
  $child->*ParentClass*parent_private_f1 = 1;
  $child->child_private_f1 = 2;
}

If parent_private_f1 is a complex type instead of a primitive, trying to expand the node in the scopes window results in the debugger sending a PHP evaluation for "$child->*ParentClass*parent_private_f" which fails in the PHP compiler and never even makes it to the evaluation phase, same if you type this string into the REPL.

This diff changes two things:

1. Removing the *ParentClass* from the object output string, resulting in valid PHP syntax being sent to the scopes request.
2. When performing an evaluation on props of a class object inside the PHP VM, if the current request is a debugger request, walk the inheritance chain and permit the debugger to deref a private property from a parent class. The debugger is allowed to violate visibility modifiers on an object class instance.

Reviewed By: paulbiss

Differential Revision: D5837175

fbshipit-source-id: bb1edbdc9aff8c921c478091661834d69a2324ea