UNSAFE_EXPR comments are treated as always leading
Summary:
We have a magical /*UNSAFE_EXPR*/ comment which is treated as an invisible operator before any expression; it causes that expression to have errors suppressed during type reporting.

The original parser actually treats such comments as operators when seen in an expression context. The FFP does not; it treats them as comments, and then the lowering engine translates them into the original format.

There are a number of problems here. The problem addressed by this diff is that

    $foo = /*UNSAFE_EXPR*/whatever;

is lexed with the trailing whitespace and comments being treated as part of the =, not part of the following expression, and the lowering code then does not apply the operation to the lowered tree.

A principled fix for this would be to revisit the entire design of the feature; if "unsafeness" is actually an operator then *make an operator*, not a comment. Something like

    $foo = unsafe(whatever);

or

    unsafe {
      $foo = whatever;
    }

would be perfectly reasonable. (See the design of the C# "unchecked" and "unsafe" features.)

Pending that principled language design fix, we address the problem here by simply changing the policy for how trivia are associated with tokens.  The previous rule was that all trivia up to AND INCLUDING the first newline trivia are the trailing trivia of the preceding token, and then the rest are the leading trivia of the following token.  The modified rule is that trivia up to but NOT INCLUDING the first UNSAFE_EXPR comment are trailing.

The lowering layer then treats the leading comment as an operator.

Note that the lowering layer has an additional problem not addressed by this change.  After this change, an expression such as

    $foo = /*UNSAFE_EXPR*/whatever + something;

will lower this to the moral equivalent of

    $foo = unsafe(unsafe(whatever) + something);

and not

    $foo = unsafe(whatever + something);

which is what was intended. This should be harmless.

 

Differential Revision: D5855548

fbshipit-source-id: 2b4124786601012587c43db9eb72d1b3df6a5a02