Experimental error for mismatched pass-by-ref annotations
Summary:
HHVM now provides `Eval.WarnOnCallByRefAnnotationMismatch`, which when set causes the runtime to raise warnings if `&` annotations on function arguments at the call site do not match the definition.

This change introduces an analogous Hack error behind an experimental feature named `safe_pass_by_ref`. When set in the `.hhconfig` or as a flag passed to `hh_single_type_check`, the type checker will report error 4168 at every argument that is misannotated.

Reviewed By: jamesjwu

Differential Revision: D5858040

fbshipit-source-id: 5e5b7cd5ae63fd1fe6eec63736b76d282c7482b9