Fix the interaction between static methods and Lvar $this
Summary: This diff fixes some interactions between static methods and Lvar `$this`. Before this diff, `Ast_scope.Scope.has_this` was providing a rudimentary way to make decisions. This diff generalizes this and exposes an API at which we can check whether we are in a static context inside a class. This enables us to correctly deal with  interactions between static methods and Lvar `$this.

Reviewed By: vladima

Differential Revision: D5857940

fbshipit-source-id: 58a5c6e4913d18b8e85657d2e186fe06e288f251