Add constraints to the environment before localizing where clauses
Summary:
When we localize a function type, we localize its type parameters independently
from the where clause. But if we access a type constant on a generic, we need
to know that the generic is upper bounded by a constraint in order to ensure
that the type constant exists on the generic. Thus, this diff temporarily
adds constraints from type parameters into the environment in order to
localize where clauses properly. Afterwards, it resets the tpenv.

it already passed all the tests before, so I'm

ing it to get it into the release

Differential Revision: D5788825

fbshipit-source-id: 064d353b6f1308d5c9f50de532fbd5bb5e6e32b3