add some more genva-like functions
Summary:
Not sure we're going to add these, but supporting them in the typechecker, even if they don't exist, doesn't hurt. If we /do/ do this in HSL v1.0, it would be good to have them supported in 3.22 (due friday)

Open question: I've made these also take variadics, rather than a tuple. This is:
 - more convenient
 - consistent with existing implementations
 - inconsistent with vec/dict/keyset

I think this is the right choice as:
 - I don't remember *ever* seeing a real `(Awaitable<T1>, Awaitable<T2>)`
 - the wrapping will annoy people :p

Context: https://www.facebook.com/groups/hhvm.general/permalink/897172530436141/?comment_id=897199327100128

Reviewed By: kmeht

Differential Revision: D5873888

fbshipit-source-id: 12755118893b7a7488f1f550720fac54a84c320b