Extend register sign using VASM for PPC64
Summary:
The VASMs ```loadw``` and ```loadl``` was implicity extending the sign when
emitting the instruction. It was creating invalid VASM sequences, as right
after a ```loadw``` may appear a ```cmpq```, which uses different register sizes.
Another inconsistency was the ```loadb``` VASM instruction, which for ```PPC64``` by
default is not signed. When loading a byte and comparing with another
value, the invalid VASM sequence ```loadb```; ```cmpq``` was being created. It was
fixed lowering the ```loadb``` instruction to ```loadzbq```, in order to explicitly
extend the size.
Closes https://github.com/facebook/hhvm/pull/7965

Differential Revision: D5727312

Pulled By: mxw

fbshipit-source-id: 337324532e8274cc135fdc9cbbca1797ddb7cc3b