Suppress type hint errors on "define" style constants.
Summary:
Due to a confluence of unfortunate error recovery behaviours, it is possible to get the type checker to report a "missing type hint" on a top-level "define('FOO', 123);" statement.

In the long run, we will probably make (1) top level FIXME, and (2) top-level defines both illegal and codemod them away. In the meantime though, I've modified the type checker so that we track whether a constant came from a define or a normal const declaration. We suppress the missing type hint error if it came from a define.

Differential Revision: D5878227

fbshipit-source-id: a14a608bf79050f5635ead43879d4646bd82e009