Create to_string helper for fixme collection
Summary: Fixmes are stored in a multi-map that maps (file name, line number, error number) to position range. Here I'm adding a helper function to dump out the line-number / error number / position table for a particular file.  I'm debugging an issue involving FIXMEs being lowered differently in the FFP and the original parser, and this will be useful.

Reviewed By: oulgen

Differential Revision: D5885231

fbshipit-source-id: 0f9ab5db54afb6459ede0dcbab6db3cdb3aed95b