Rename State_printer to Subchunk
Summary:
This change renames State_printer to Subchunk, clarifying that that module is about the subchunk data structure, converting solve states to subchunks, and converting subchunks to strings.

The print_state function is also refactored to return an unfiltered list of subchunks, so that range filtering can happen in a single place.

Reviewed By: arxanas

Differential Revision: D5843287

fbshipit-source-id: bd4faa80c41a2b805e1f03a8fbfc9556bbc9a6a5