Expand to atom boundaries rather than split boundaries in diff and at-offset modes
Summary:
Currently, we expand intervals in diff mode and at-offset mode to line boundaries in the original source, then split boundaries in the formatted source. This is usually correct, but rarely, split boundaries in the formatted source won't correspond to line boundaries in the original source.

Instead, this change expands the formatting range to the nearest outer atom boundaries, then formats that range in a mode which does not produce any leading or trailing whitespace trivia. This should always produce the correct result, since we are expanding the range to include all surrounding whitespace (which ought to be considered hackfmt-managed), instead of replacing only *some* of the whitespace and hoping that the surrounding whitespace in the formatted string is correct.

Reviewed By: arxanas

Differential Revision: D5843678

fbshipit-source-id: 0dd6cfb98cd3644a7c874502fb818e01e5f6a2cf