Change chunk ranges to return the span of their atoms
Summary:
Originally, the start_char-end_char range of a Chunk was the only source mapping signal used for range formatting. Unlike atom ranges, these ranges make up a complete mapping from characters in the original source to chunks (that is, the end_char of one chunk is always equal to the start_char of the next chunk).

This sounds nice, but this mapping is created by handling trivia at the same time that we handle tokens (a sensible thing to do, considering that trivia is associated with tokens in the full fidelity syntax tree). This means that the mapping is incorrect for XHP. Since the lexer handles XHP body trivia in an unusual way, chunk ranges are off by one in XHP expressions.

Instead of trying to fix chunk ranges to map to the original source more reliably, this change removes our reliance on chunk ranges altogether. Now that we don't expand ranges to split boundaries, the only dependency on chunk ranges is Solve_state.from_source, which now scans between atoms for newline characters in the original source instead of relying on a Chunk's start_char accurately pointing to the character after a newline character.

Differential Revision: D5848656

fbshipit-source-id: 1c4e7847c4cc3e2652afac1ff4d0345d6eaf1f3c