add utility module for RSTDC counter
Summary:
RSTDC gives you access to the processor clock, which can be useful if you're
trying to benchmark something small.

A 64-bit float should have reasonable precision over an unsigned 64 bit integer's range.

I'm not sure what to do in order to handle the case where the program is being compiled on a platform that doesn't support this instruction.
As of right now it's a compile-time error.

Reviewed By: michaeltingley

Differential Revision: D5885411

fbshipit-source-id: 5ea6e74a32297f7dc0098e5cd60638f275fbd0f2