Keep NewVecArray and NewKeysetArray from using too much stack
Summary:
Limit the amount of stack space that NewVecArray and NewKeysetArray can use, in
a manner similar to to NewPackedArray. This keeps very large vec and keysets,
which aren't static, from causing quadratic time blow-ups in the compiler.

Make sure HHBBC honors the size threshold.

Previously, if the array went over the size threshold, we would not use
NewPackedArray at all and generate the array element by element. Instead, use
NewPackedArray on the size below the threshold, and use AddNewElemC for the
rest. This required implementing AddNewElemC for vecs and keysets.

Bump up the size threshold to 64 to match the max struct array size. The struct
array size limit is for the exact same reason, so it makes no sense to have
different values for them. Consolidate all the constants into a single one.

Modify AddNewElemC to dec-ref the appended value in the IR instead of in the
runtime helper, which should expose it better to ref-count optimizations. Even
better would be to just teleport it from the stack, but there's no API to do
that right now.

Remove NewMIArray and NewMSArray from the documentation, as they haven't existed
for a long time.

Reviewed By: markw65

Differential Revision: D5892079

fbshipit-source-id: b1d7004be3fc98c4184fbc6ac85f5cfc8d67b55c