allow completion to insert invocation-snippet in more situations
Summary:
Previous behavior of Nuclide+Hack was this: "If the user presses tab to autocomplete a method then we'd like to autocomplete it with parentheses and placeholder arguments. EXCEPT: if there were already characters after the caret, and they were zero or more whitespaces followed by an open-parent, then we don't want to autocomplete with parentheses."

Typical scenario: `if ($q->|)` we want to auto-insert the invocation `$q->foo()` in here.

HackLSP doesn't do this. HackLSP has a worse algorithm which says "... EXCEPT: only use parentheses+placeholders if the character after the caret is a whitespace".

This diff does something a bit simpler than the previous behavior. This diff says "... EXCEPT: don't use parentheses+placeholders if the character after the caret is an lparen".

This was a quick and easy fix which I imagine will satisfy most scenarios. It arguably won't do the right thing in the case `$q->| ()` where parentheses already existed but after a whitespace. I write "arguably" because it's not clear what the right thing to do is.

Differential Revision: D5884514

fbshipit-source-id: 3cd8e0aa7791b71135c53b44e8b738bd5538503d