go-to-def should go directly to constructor
Summary:
What should go-to-def do when you click on a "new Xyz()" construction?

hh_server actually returns two definitions for this, one to the class, and one to the constructor. But it's kind of clunky to surface both of these to the user. Is much nicer to send the user directly to the constructor.

We tried to do this in an earlier diff, but didn't get it quite right:
https://github.com/facebook/hhvm/commit/e66bdf4bae19344fb329125a09d1e2510bfa7262

What happens: when we query hh_server for the definition, it actually gives two results:

1. Result1 has Occurrence.Kind=Class and points to the class declaration (i.e. Definition.Type=Class)
2. Result2 has Occurrence.Kind=Method and points either to the constructor declaration Definition.Type=Method (if there was one) or to the class declaration Definition.Type=Class (if there wasn't).

The logic used in the earlier diff was that, if there were multiple results and one of them pointed to the class declaration and the other pointed to the constructor declaration, then pick only the constructor declaration.

That fails in cases where there was no explicit constructor declaration\!

This diff fixes it to be bassed on Occurrence.kind instead.

Reviewed By: arxanas

Differential Revision: D5858164

fbshipit-source-id: 0ca81611b166c85d814802fdf5be88c00bf8f0c0