add typechecking for suspend operator and coroutines
Summary:
Step 3 of the efforts to enable typechecking of coroutines.

This diff adds typechecking of `suspend` operator and coroutine functions.
Currently coroutine functions in Hack are represented as regular functions with a annotated with `ft_is_coroutine=true`. From the typechecking perspective coroutine functions and regular functions are considered different incompatible types but other than that rules for typechecking coroutine functions and regular functions are pretty much the same with a few exceptions:
- `suspend` operator is used to denote suspension points and can only appear inside coroutine functions
- `suspend` operator can only accept calls to coroutine functions as its argument. Also this is the only place where calls to coroutine functions can appear in the program, it is illegal to call the in any other places
- if `f` is a coroutine function with a return type `T` and result type of the expression `suspend f()` is `T`.
-

Reviewed By: michaeltingley, andrewjkennedy

Differential Revision: D5616762

fbshipit-source-id: 11e1e86bb11e43ae8368c610470b80f93447e90b