Turn FFP errors upto 11; report everything
Summary:
So far, the lowerer for the Full Fidelity Parser propagated errors quietly. If the use site didn't ask for them, they would go unnoticed. To match the behaviour of `Parser_hack`, when `keep_errors` is set, no AST should be produced if there are parse errors.

This diff invokes `Full_fidelity_parser_errors.parse_errors` on the error at the bottom of the error stack (this was produced first). Since the desired behaviour is to bail without an AST, this is done prior to lowering.

Reviewed By: andrewjkennedy, hubyrod

Differential Revision: D5697549

fbshipit-source-id: 89aa9bd51e96df8a4bcd5cc8ef5f97d8531b65cd