Handling `UNSAFE` and `UNSAFE_EXPR` in FFP/lowerer
Summary:
The Full Fidelity Parser's lowerer didn't yet deal with `UNSAFE` and `UNSAFE_EXPR` pragmas. The FFP already correctly detects and marks trivia as containing these pragmas (see, for example, `TriviaKind.Unsafe`).

This diff adds support for this in the lowerer. The main problem is that the FFP prefers to put comments in a token's *trailing* trivia, which means that in
```
return /* UNSAFE_EXPR */ $x;
```
the pragma is part of the `Return` token's trailing trivia, rather than the `Variable` token's leading trivia. This makes the lowering a bit ad hoc, but testing should figure out whether we catch them all, before we turn these pragma's into language-level constructs, with unambiguous scopes.

Reviewed By: hubyrod

Differential Revision: D5697565

fbshipit-source-id: 761234094f42e3c2ef7d9721ab9b8fab0ef9226b